+++ src/components/TmLiTransaction/index.js
@@ -1,5 +1,5 @@
 import TmLiTransaction from "./TmLiTransaction.vue"
 
 export default function install(Vue) {
-  Vue.component("tm-li-tx", TmLiTransaction)
+  Vue.component("tm-li-transaction", TmLiTransaction)
 }